twitch/pyproject.toml
[project]
name = "twitch-crayon"
version = "0.1.0"
description = "A simple Twitch bot to connect to home appliances"
authors = [{ name = "Tehoor Marjan", email = "tehoor.marjan@gmail.com" }]
dependencies = [
    "twitchAPI>=4.5.0",
    "click>=8.3.1",
    "platformdirs>=4.5.1",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[project.scripts]
tc = "twitch_crayon.cli:main"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "black>=25.12.0",
    "isort>=7.0.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "jinja2>=3.1.6",
    "ipywidgets>=8.1.8",
    "bs4>=0.0.2",
]
